refactor: a single tensor send/recv (#43)
Code is refactored so that we only send/receive one tensor via api, not a tuple of
tensors. The change is to ensure that model expects only one tensor input
(i.e. requests, prompts, etc). To make this change work, sharding code is refactored
such that only one extra output is added to the node, not as tuple.