fix: comm error fix via grpc (#27)
When controller and agent communicate with each other, there is a serialization
error. This is because the controller's servicer code doesn't return
empty_pb2.Empty(). In addition, including_default_value_fields is deprecated
since v5.26, which also raises an error. These errors are now fixed.
Since a new version of protobuf is used, protobuf code was regenerated; note
that the staleness of the autogenerated protobuf code has nothing to do with
the problem.