feat: llama3 distirbuted serving (#39)
serving llama3 in a distributed fashion is now supported.
Because of the autoregressive nature of token generation in llm and
lack of admission control in infscale, infscale continues to accept
other requests while tokens keep generated for earlier requests.
This results in a lengthy response time. So, a simple admission control
logic is implemented at a server (dispatcher) side by introducing
the maximum number of requests that can be in flight at any given point
in time. That led to introduction of a new variable in the config file.