refactor: preparation of router for online reconfiguration (#47)
To enable online reconfiguration of pipeline, router is refactored.
Now router is initialized only with static elements. Since the elements
in a config can be updated, those are passed as variables through
configure method.

Later, router will maintain this dynamic information so that it can
reconfigure whenever the config is updated. This will be handled in
future PRs.