feat: implemented stopped condition for a job (#81)
Refactored the way messages were passed between worker communicator and worker manager. Before, whenever a terminate message was received from the agent, we were closing the pipe communication.
Since workers need to do some clean-up before setting the actual terminated status, they had to send back a status message with terminated in order to let the agent know that they were successfully terminated.
Therefore, right before killing the worker process, an terminated status message is sent to the agent to pass that status to the controller.
Based on these changes, since the controller receives worker status messages, job state checks if all the workers are terminated for a specific job.
If all workers are terminated and the current job state is stopping, then the state can be transitioned to stopped.