fix: ccl timeout (#100)
When worker initialization takes too long due to model loading and so on,
its ccl initialization (connection) can fail. The default connection timeout
value is 30 seconds. Since there is no gauarantee that worker initialization
can be done in 30 seconds, we make sure that connection delay can be up to
5 minutes. In doing so, we swap control channel setup and multiworld (ccl)
channel setup. We increase the control channel setup timeout to 5 minutes.
This is still a temporary fix since 5 minutes can be insufficient too.
Implementation on a permanent solution is deferred as a future task.