feat: implemented worker distribution state logic for agent data (#107)
Added worker distribution to each AgentMetaData to be able to keep track on the workers that are being deployed by each agent.
Code was refactored in such way that the deployment policy will handle the agent metadata update, based on the split result.
With these changes, that issue with updating a job that was deployed in multiple agents is fixed, since we now have a reference on how the workers are distributed between agents.