feat: implemented device assignment when split is called (#123)
Split method will receive agent resources class instances to properly assign device and update agent resources afterwards.
Agent resources class was updated to expose a new method for deployment policy to use.
Also, worker distribution signature was updated to handle pair of worker ID and device.
GPU is the preferred device in this PR, but more work is needed to decide how to update CPU resources if a worker is deployed on CPU and also to check if enough resources are available on current agents to deploy the workers.
Code refactor to improve readability and reduce computation was added in job context