feat: exception format handler (#142)
Since we need to validate incoming requests in the controller, a new exception format handler was implemented when receiving FastAPI requests.
If RequestValidationError is dispatched by FastAPI, we map this request and we create a user friendly message to be returned to the client.
Also, added this custom exception handler to FastAPI app.