feat: exponential request generator (#174)
Requests are generated based on an exponential distribution.
The rate parameter in the exponential distribution controls an
inter-arrival time of requests. The rate's unit is batch/second.
To choose request generators flexibly, controller now takes request
generator config as an argument. Since requests can arrive in a bursty
fashion, generator returns a list of requests (tensors) rather than a
single request. To accommodate this change, pipeline code is refactored.