feat: autoscale (#225)
autoscaling functionality is implemented. Various refinements were
made. (1) congestion detection mechanism is simplified as the
input and output rate changes are unstable. (2) if the current and
new configs are identical, there is no need to deploy the new config.
Detecting identical configs are enabled. (3) To ensure incremental
job update (or autoscaling), cfggen code is revised.