feat: delta autoscale (#234)
The current autoscale functionality creates an entire new configuration
to deal with the total demand. However, this implementation complicates
the management of GPUs already used for the job, which is handled by
introducing job_id in the GpuStat data structure.

To make the implementation less complex and incremental autoscaling easy,
a new config to handle the additional demand is created and merged with
the existing config.