refactor: conditional worker fail (#284)
Treating multiworld and channel connection as a fatal failure is wrong in certain context. Whenever we have an ongoing update, killing worker when these processes fail is incorrect, because we can always roll back to previously running config. For that reason, we need to properly store worker status and check against that in order to decide terminating worker process.