fix: pending requests not forwarded on worker termination (#316)
When a new config is received that is supposed to terminate some workers, we blindly close connection between these workers and their receivers, which blocks the worker to be terminated forward remaining requests to it's receiver. In order to fix this, code was updated in a way that only the senders of the workers to be terminated will close their connection and the receivers of the workers to be terminated will close the connection when a broken world exception is raised, meaning that the worker is terminated and all requests were forwarded. Also, refactored the code so the cleanup is done in router only.